refactor: applications sort order 관련 파일 이동
